Move general test cases into the `runtime` package.
